+++ enclone/src/lib.rs
@@ -12,6 +12,7 @@ pub mod innate;
 pub mod join;
 pub mod join2;
 pub mod join_core;
+pub mod join_one;
 pub mod join_utils;
 pub mod load_gex;
 pub mod misc1;
